pyproject.toml
[project]
name = "dcbs-eval"
version = "0.1.0"
description = "Deterministic Category Based Sampling evaluation harness"
authors = [{name = "Aviad Cohen"}]
dependencies = [
    "torch>=2.3.0",
    "transformers>=4.43.0",
    "tokenizers>=0.13.3",
    "accelerate",
    "bitsandbytes",
    "safetensors", 
    "sentencepiece",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "tabulate",
    "requests",
    "huggingface-hub>=0.15.0",
    "jinja2>=3.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "flake8-black>=0.3.0",
    "flake8-isort>=6.0.0",
    "isort>=5.12.0"
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dcbs", "src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

# Code formatting with Black
[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Import sorting with isort  
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["dcbs", "src"]
skip_gitignore = true

# Testing with pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 